requirements.txt
Django==5.0.2
django-crontab==0.7.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
